+++ src/main/java/org/datadog/jmxfetch/JmxFetch.java
@@ -46,8 +46,9 @@ public class JmxFetch {
 
             // Set up the logger to add file handler
             CustomLogger.setup(LogLevel.fromString(config.getLogLevel()),
-                    config.getLogLocation(),
-                    config.isLogFormatRfc3339());
+                config.getLogLocation(),
+                config.isLogFormatRfc3339(),
+                config.isLogThreadName());
 
             // Set up the shutdown hook to properly close resources
             attachShutdownHook();